* Fix benchmarks * Add staking_history to StakingAccountDetails * Change type of some Configs * Stub some tests * separate provider_boost tests into own file
