+++ pallets/capacity/src/tests/stake_and_deposit_tests.rs
@@ -51,70 +51,6 @@ fn stake_max_capacity_works() {
 }
 
 #[test]
-fn provider_boost_works() {
-	new_test_ext().execute_with(|| {
-		let account = 600;
-		let target: MessageSourceId = 1;
-		let amount = 200;
-		let capacity = 1;
-		register_provider(target, String::from("Foo"));
-		assert_ok!(Capacity::provider_boost(RuntimeOrigin::signed(account), target, amount));
-
-		// Check that StakingAccountLedger is updated.
-		let staking_account: StakingAccountDetails<Test> =
-			Capacity::get_staking_account_for(account).unwrap();
-
-		assert_eq!(staking_account.total, 200);
-		assert_eq!(staking_account.active, 200);
-		assert_eq!(staking_account.unlocking.len(), 0);
-		assert_eq!(staking_account.last_rewards_claimed_at, None);
-		assert_eq!(staking_account.stake_change_unlocking.len(), 0);
-
-		let events = staking_events();
-		assert_eq!(
-			events.first().unwrap(),
-			&Event::ProviderBoosted { account, target, amount, capacity }
-		);
-
-		assert_eq!(Balances::locks(&account)[0].amount, amount);
-		assert_eq!(Balances::locks(&account)[0].reasons, WithdrawReasons::all().into());
-
-		let target_details = Capacity::get_target_for(account, target).unwrap();
-		assert_eq!(target_details.amount, amount);
-		assert_eq!(target_details.staking_type, ProviderBoost);
-	});
-}
-
-#[test]
-fn calling_stake_on_provider_boost_target_errors() {
-	new_test_ext().execute_with(|| {
-		let account = 600;
-		let target: MessageSourceId = 1;
-		let amount = 200;
-		register_provider(target, String::from("Bear"));
-		assert_ok!(Capacity::provider_boost(RuntimeOrigin::signed(account), target, amount));
-		assert_noop!(
-			Capacity::stake(RuntimeOrigin::signed(account), target, 50),
-			Error::<Test>::CannotChangeStakingType
-		);
-	})
-}
-#[test]
-fn calling_provider_boost_on_staked_target_errors() {
-	new_test_ext().execute_with(|| {
-		let account = 600;
-		let target: MessageSourceId = 1;
-		let amount = 200;
-		register_provider(target, String::from("Foobear"));
-		assert_ok!(Capacity::stake(RuntimeOrigin::signed(account), target, amount));
-		assert_noop!(
-			Capacity::provider_boost(RuntimeOrigin::signed(account), target, 50),
-			Error::<Test>::CannotChangeStakingType
-		);
-	})
-}
-
-#[test]
 fn stake_errors_invalid_target_when_target_is_not_registered_provider() {
 	new_test_ext().execute_with(|| {
 		let account = 100;
